Add generic peer-to-peer transfer connector
- Add generic P2PXferConnector
- Refactor NixlConnector to extend P2PXferConnector
