+++ vllm/envs.py
@@ -161,6 +161,9 @@ if TYPE_CHECKING:
     VLLM_XGRAMMAR_CACHE_MB: int = 0
     VLLM_MSGPACK_ZERO_COPY_THRESHOLD: int = 256
     VLLM_ALLOW_INSECURE_SERIALIZATION: bool = False
+    VLLM_P2PXFER_SIDE_CHANNEL_HOST: str = "localhost"
+    VLLM_P2PXFER_SIDE_CHANNEL_PORT: int = 5557
+    VLLM_P2PXFER_ABORT_REQUEST_TIMEOUT: int = 120
     VLLM_NIXL_SIDE_CHANNEL_HOST: str = "localhost"
     VLLM_NIXL_SIDE_CHANNEL_PORT: int = 5600
     VLLM_ALL2ALL_BACKEND: Literal[
@@ -1191,6 +1194,18 @@ environment_variables: dict[str, Callable[[], Any]] = {
     "VLLM_ALLOW_INSECURE_SERIALIZATION": lambda: bool(
         int(os.getenv("VLLM_ALLOW_INSECURE_SERIALIZATION", "0"))
     ),
+    # IP address used for P2P XFER handshake between remote agents.
+    "VLLM_P2PXFER_SIDE_CHANNEL_HOST":
+    lambda: os.getenv("VLLM_P2PXFER_SIDE_CHANNEL_HOST", "localhost"),
+    # Port used for P2P XFER handshake between remote agents.
+    "VLLM_P2PXFER_SIDE_CHANNEL_PORT": lambda: int(
+        os.getenv("VLLM_P2PXFER_SIDE_CHANNEL_PORT", "5557")),
+    # Time (in seconds) after which the KV cache on the producer side is
+    # automatically cleared if no READ notification is received from the
+    # consumer. This is only applicable when using P2PXferConnector in a
+    # disaggregated decode-prefill setup.
+    "VLLM_P2PXFER_ABORT_REQUEST_TIMEOUT": lambda: int(
+        os.getenv("VLLM_P2PXFER_ABORT_REQUEST_TIMEOUT", "120")),
     # IP address used for NIXL handshake between remote agents.
     "VLLM_NIXL_SIDE_CHANNEL_HOST": lambda: os.getenv(
         "VLLM_NIXL_SIDE_CHANNEL_HOST", "localhost"
